added controllers and seperate routing files. Required routes folder to serve all the requests
