+++ server.js
@@ -3,7 +3,7 @@ const PORT = process.env.PORT || 3001;
 const app = express();
 const mongoose = require("mongoose");
 const passport = require("passport");
-const users = require("./routes/api/users");
+const routes = require("./routes");
 
 const bodyParser = require("body-parser");
 const upload = require("./routes/api/image-upload");
@@ -38,8 +38,8 @@ app.use(passport.initialize());
 require("./config/passport")(passport);
 
 //Routes 
-app.use("/api/users", users);
-app.use("api/upload",upload);
+app.use(routes);
+//app.use("api/upload",upload);
 
 app.listen(PORT, () => {
   console.log(`🌎 ==> API server now on port ${PORT}!`);
